requirements.txt
Pillow==2.3.0
pyserial==2.7
